Add files via upload
Several changes were applied to the CSS and disposition of the page. New CSS projects uploaded according to my learning process. This is an update to keep the concurrency since I am still learning flexbox and bootstrap and I will be updating the page shortly with a new fresh disposition.